# openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityRefItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CapabilityListResponse",
        "type": "object"
      },
...
      "CreateCapabilityRequest": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "capability_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Type"
          },
          "entity_type": {
            "const": "capability",
            "default": "capability",
            "title": "Entity Type",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateCapabilityRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Paste an API token.\n\n- **Local development:** run `make dev-token` \u2014 it seeds a tenant + actor + role grants and prints a fresh token. No env-var setup required if the docker-compose stack is up.\n- **Production / break-glass:** mint with `python scripts/mint_token.py --tenant-id <uuid> --actor-id <uuid> --roles <role>`. Requires the tenant and actor to already exist.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "description": "Paginated list of capabilities visible to the caller's tenant.\n\nPagination is keyset-based: the response carries ``next_cursor`` (or null\nwhen no further pages exist). Pass ``cursor=<value>`` on the next request\nto retrieve the following page.\n\nThe legacy ``?page=N`` offset parameter is no longer accepted. Clients\nthat send it receive a 422 with code ``page_param_deprecated``.",
        "operationId": "list_capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "description": "Opaque cursor returned by the previous page. Omit to start from the first page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor returned by the previous page. Omit to start from the first page.",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "ISO-8601 UTC datetime for time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC datetime for time-travel",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "retrieval"
        ]
      },
      "post": {
        "description": "Create a new capability.\n\nHonours ``X-Idempotency-Key`` (optional). Resend with the same key\n+ same body \u2192 returns the original response. Same key + different\nbody \u2192 409 with ``code: \"idempotency_key_conflict\"``.",
        "operationId": "create_capability_v1_capabilities_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCapabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Capability",
        "tags": [
          "capabilities"
        ]
      }
    },
    "/v1/capabilities/{capability_id}/interface": {
      "get": {
        "description": "Return the active interface surface at ``as_of`` (or current truth).\n\nThe path segment accepts a UUID or slug-form name.\n\n``?view=audit`` is accepted for API consistency but is currently a no-op \u2014\nthe interface service returns a composed record rather than raw attribute\nrows, so no additional bitemporal metadata is available to surface.",
        "operationId": "get_interface_v1_capabilities__capability_id__interface_get",
        "parameters": [
          {
            "description": "Capability UUID or slug",
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "description": "Capability UUID or slug",
              "title": "Capability Id",
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 UTC for time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC for time-travel",
              "title": "As Of"
            }
          },
          {
            "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` is accepted for API consistency but is currently a no-op here \u2014 the interface record is a composed view without individual bitemporal row metadata. Use ``?as_of=`` for time-travel instead.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` is accepted for API consistency but is currently a no-op here \u2014 the interface record is a composed view without individual bitemporal row metadata. Use ``?as_of=`` for time-travel instead.",
              "title": "View",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the capability's declared interface surface",
        "tags": [
          "interface"
        ]
      },
      "put": {
        "description": "Normalize, soft-supersede prior versions, then write the new pair.\n\nThe path segment accepts a UUID or slug-form name.",
        "operationId": "put_interface_v1_capabilities__capability_id__interface_put",
        "parameters": [
          {
            "description": "Capability UUID or slug",
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "description": "Capability UUID or slug",
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfacePutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceSurfaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the capability's declared interface surface",
        "tags": [
          "interface"
        ]
      }
    },
...
    "/v1/capabilities/{entity_id}/dependencies": {
      "get": {
        "description": "k-hop dependency traversal from entity_id.\n\nPath segment accepts UUID or slug-form name. Depth capped at 5 by\nthe service layer.",
        "operationId": "get_dependencies_v1_capabilities__entity_id__dependencies_get",
        "parameters": [
          {
            "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 5,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "description": "ISO-8601 UTC datetime for time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC datetime for time-travel",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dependencies",
        "tags": [
          "retrieval"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}